Suggestions for PR #1611
